Motor sliders now update every second (instead of on focus).
Per the modified implementation of ticket 8, we now update the motor
sliders regularly (so they more precisely match reality). As an
interesting side-effect, it makes more plain the issue of ticket 3
(rotational motor hiccup): degrees that share a two-radian value with
another degree are 'difficult' to go past (since the motor resets to one
of the two possible degrees). This will likely become my next task. :)

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>