First revision of SPIM calibration class.
It has yet to do anything supremely valuable, but in the future I have
hopes that it will provide for (first manual, hopefully shortly followed
by) automatic determination of the axis of rotation, as well as a method
of choosing the pixel size graphically. This will help the live window
controls react appropriately, as well as let us provide some accurate
metadata for captured images.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>