Massive upgrades to (semi-)auto calibrator.
It now looks like something I might eventually be able to convince myself
to release. The detection could use some work yet, and I haven't found a
good way to go around a huge chunk of the circle -- preferably well over a
quarter -- but those are minor issues to be fixed later.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>