Modified ProgAcq to better understand stacks
Also hid some of the dirt from the user -- there's probably quite a few
changes to be made yet; I'd like to eventually replace rows of strings
with a StepRow class or the like, to better describe the steps to be
taken. Also, there's probably some unsafe things going on that I'll need
to track down shortly, but pushing this out was a bit of an emergency.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>