Stopped querying laser while acquiring
Unfortunately, the Cube DAL doesn't sequence its queries; if a query
occurs before the last is responded to, a timeout error occurs (because
the receive buffer is now empty). I'm considering fixing this in the code,
but for the purposes of acquisition, stopping the queries to the laser is
sufficient.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>