Refactored anti-drift for easier testing/changing
I've pulled all the anti-drift code into a separate class which is now
used. It also is a completely different implementation adjusting based on
each previous stack, so we don't need to take each stack twice.

It does, however, have a serious issue: in the live tests, the
measurements tend to jump out of control; Z in particular oscillates
between positive and negative while its amplitude grows. I believe this to
be a logical mistake somewhere; I'm working on tracking it down.

Additionally, the background subtraction at the moment is a bit shady. I'm
using the max/min properties of ImageProcessor during each frame to pull
out that slice's unique background (though the code is a single
coefficient away from using an overall background subtraction). It seems
to be causing unusual results, which I'll be testing shortly.

A particular advantage of this new implementation is that we can use e.g.
Beanshell to test it on arbitrary stacks. This will let me develop and
refine it far more quickly than before.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>