First version of asynchronous output
I know I pushed it to milestone 2 on the bugtracker, but I think these
results are pretty convincing of its necessity:

100 slices per view:
1 view, 1 timepoint, synchronous: 39.188s total, 24.559s writing (63%)
1 view, 1 timepoint, asycnhronous: 22.485s total, 6.820s writing (30%)
4 views, 4 timepoints, synchronous: 617.865s total, 392.188s writing (63%)
4 views, 4 timepoints, asynchronous: 334.638s total, 96.014s writing (29%)

As you can see, asynchronous writing has nearly halved the time required
to record images. The wrapper is simple and can be applied to
(theoretically) any output handler we develop in the future.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>