Added acquisition options dialog
The more obscure/less used options will be relegated to a dialog
accessible by a button on the acquisition tab. This lets me clean up the
UI by removing lesser-used options from the main window, and also put in
options for debugging that won't bother the users. For example, there is
now a checkbox under that dialog which enables/disables the async output
monitor; most users may not care what the async is presently doing, but if
it's being problematic for someone, they can check on it that way. I
intend to add acquisition profiling to this dialog as well.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>