Revamped live controls
The live-window mouse controls are now a separate class. This lets me add
some useful functionality and also fixes a bug where, after dragging in
the live window, then moving the stage via the sliders, the stages would
move to their pre-drag positions.

The new controls, however, do not automatically hook -- there is now a
checkbox on the bottom of the first tab to enable them, though they will
warn about the presence of MM's live window hooks. Rotation via the mouse
will not be available until the checkbox is manually clicked.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>