TEMP: Don't automatically open live view
A change upstream flipped some part of the load order, and now the
original live window controls aren't created until after the plugins are
loaded and the autostart script is run -- which means that our opening the
MM live window on startup causes an exception that disables mouse controls
completely. For the moment, don't open live view automatically.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>