Add automatic peak detection to auto-calibrator
Please note that this commit makes use of a version of Imglib2 which is
not at present released on the Fiji update site; as a result, it shouldn't
be included in the OpenSPIM plugin just quite yet.

The auto-calibrator will pick out the brightest bead when it's opened (or
when a button under 'tweaks' is clicked). This will hopefully lead to a
quicker calibration process.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>