Upgrade to new Micro-Manager
Despite having a hackathon with the Micro-Manager developers in Madison
in July 2014, we failed to merge any code of the OpenSPIM fork back into
the Micro-Manager code base.

Lots of updates went into Micro-Manager after the hackathon, and if we
had bothered to contribute our plugin upstream, the Micro-Manager team
would have adapted the source code accordingly, together with all the
other plugins that needed adjusting anyway.

This is unfortunate, but cannot be helped. The only point of this lament
is to try to make it unlikely that such a mistake be repeated.

To compile this, you need to install both MMJ_ and MMCoreJ, say, from a
current nightly build via:

	mvn install:install-file \
		-DgroupId=org.micromanager -Dversion=1.4.19-SNAPSHOT \
		-Dpackaging=jar -DartifactId=MMJ_ -Dfile=.../MMJ_.jar

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>