Fix XY stage movement after homing
Peter Gabriel Pitrone pointed out that after homing the 4D stage,
pressing the "+" button on the x position would reset the y position to
the one *before* homing.

The culprit is that redundant information is held in too many places
(here: the MM core *and* the GenericXYStage wrapper) and such
information is prone to become inconsistent, even for the best
programmers amongst us.

While the workaround introduced by this commit is far from perfect, it
allows Peter to use the software, at least, while HongKee is working on
cleaning up the software design.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>